Add tests for id joins, reduce dependencies for tests
