+++ packages/back-end/src/services/stats.ts
@@ -13,7 +13,7 @@ import {
 import { QueryMap } from "./queries";
 import { getMetricsByOrganization } from "../models/MetricModel";
 import { promiseAllChunks } from "../util/promise";
-import { jStat } from "jstat";
+import { checkSrm } from "../util/stats";
 
 export const MAX_DIMENSIONS = 20;
 
@@ -273,30 +273,4 @@ export async function analyzeExperimentResults(
     unknownVariations: Array.from(new Set(unknownVariations)),
     dimensions,
   };
-}
-
-export function checkSrm(users: number[], weights: number[]) {
-  // Skip variations with weight=0 or users=0
-  const data: [number, number][] = [];
-  let totalUsers = 0;
-  let totalWeight = 0;
-  for (let i = 0; i < weights.length; i++) {
-    if (!weights[i] || !users[i]) continue;
-    data.push([users[i], weights[i]]);
-    totalUsers += users[i];
-    totalWeight += weights[i];
-  }
-
-  // Skip SRM calculation if there aren't enough valid variations
-  if (data.length < 2) {
-    return 1;
-  }
-
-  // Calculate and return SRM p-value using a ChiSquare test
-  let x = 0;
-  data.forEach(([o, e]) => {
-    e = (e / totalWeight) * totalUsers;
-    x += Math.pow(o - e, 2) / e;
-  });
-  return 1 - jStat.chisquare.cdf(x, data.length - 1);
 }
